+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.3] - 2019-02-22
+#### Added
+ * Requirements file
+ * .pyup.yaml file for security dependencies bot
+ 
+#### Fixed
+ * Logging when send data in multiline or zipped
+ * Error of for_logging() function in Sender when applicate level
+
 ## [2.1.2] - 2019-02-12
 #### Added
  * More documentation info and examples